requirements.txt
django_filter==2.2.0
requests==2.23.0
Django==3.0.5
django_filters==0.2.1
djangorestframework==3.11.0
